vim / neovim stuff.
I started from scratch, and found that I could make the configs here a
bit smoother. Is this important? Not so much. Was it fun? Yep.

This also removes a bunch of things, including LSP things. The landscape
here has changed. I still love vim and nvim, but my experience has been
that Visual Studio Code and other systems have stronger integrations. I
want my *vim stuff to be as fast and simple as possible, and the other
IDEs to do the heavy IDE things. Scripting small actions on buffers is
also entirely still cool, within *vim.

Maybe I'll go back on this. We'll find out. git history is wonderful.